Fix bug with directives with variables
Given something directive like,

  proxy_set_header host host2;

The lexer would would see

  DIRECTIVE_VALUE VALUE_WHITE_SPACE DIRECTIVE_VALUE

But for a directive with a variable

  proxy_set_header host $host2;

The lexer would would see

  DIRECTIVE_VALUE INNER_VARIABLE

NginxParser would create two COMPLEX_VALUEs is in the former case,
and one in the latter. This would break in NginxAnnotatingVisitor
where the call to getValues would return only one COMPLEX_VALUE,
instead of the expected two.

Error highlighting was changed to show errors on the parameters if
there is an insufficient number, else it would show it on the
directive itself.