Fix incorrect number of arguments error
When calling keywords.getValueRange, the method woould return the
first range in the set, not the most relavent one. For example,
if the ranges are (1,1) and (1, max), (1,1) would be returned
despite allowing virtually any number of arguments.

The first check is to see if there is a range that corresponds to
the given number of arguments. If there is then no error, otherwise
find the least specific range to error against.