Partial fix for Lua parsing with braces
A lua block would close at the first occurance of a closing brace,
which would brace if there were dictionaries in the Lua code.

The work around is to stop the lua context when balanced matching
closing brace is found. This only effects dictionaries, braces inside
of strings are parsed correctly. The downside of this approach is that
if you start an opening brace, the Lua context will extend far below
the actual lua block. But since it's invalid Lua it doesn't matter.